apps/backend/tests/test_middleware.py
"""
Unit tests for middleware functionality.
Tests ETag, If-Match, tracing, and logging utilities.
Does not require database connection.
"""
import pytest


class TestETagMiddleware:
    """Test ETag middleware functionality."""
    
    def test_etag_generation(self):
        """Test ETag generation from version field."""
        from app.core.etag import generate_etag, parse_etag
        
        etag = generate_etag(1)
        assert etag == '"v1"'
        
        etag = generate_etag(42)
        assert etag == '"v42"'
    
    def test_etag_parsing(self):
        """Test ETag parsing."""
        from app.core.etag import parse_etag
        
        version = parse_etag('"v1"')
        assert version == 1
        
        version = parse_etag('"v42"')
        assert version == 42
    
    def test_invalid_etag_parsing(self):
        """Test invalid ETag parsing raises error."""
        from app.core.etag import parse_etag
        
        with pytest.raises(ValueError):
            parse_etag("invalid")


class TestIfMatchMiddleware:
    """Test If-Match middleware functionality."""
    
    def test_patch_without_if_match_returns_428(self):
        """Test PATCH without If-Match returns 428."""
        # This will fail until we have a PATCH endpoint
        # For now, test the middleware logic
        pass
    
    def test_delete_without_if_match_returns_428(self):
        """Test DELETE without If-Match returns 428."""
        # This will fail until we have a DELETE endpoint
        pass
    
    def test_post_with_idempotency_key_skips_if_match(self):
        """Test POST with Idempotency-Key doesn't require If-Match."""
        # This will be tested with actual endpoints
        pass


# Integration tests moved to test_api_integration.py to avoid DB initialization issues


class TestIdempotency:
    """Test idempotency functionality."""
    
    def test_idempotency_key_extraction(self):
        """Test idempotency key extraction from headers."""
        from app.core.idempotency import IdempotencyManager
        from fastapi import Request
        from starlette.datastructures import Headers
        
        # Mock request with idempotency key
        headers = Headers({"Idempotency-Key": "test-key-123"})
        
        # This is a simplified test - in real scenario, use actual Request object
        assert "Idempotency-Key" in headers
    
    def test_request_hash_computation(self):
        """Test request hash computation."""
        from app.core.idempotency import IdempotencyManager
        
        body1 = {"product_id": "123", "quantity": 2}
        body2 = {"quantity": 2, "product_id": "123"}  # Same data, different order
        body3 = {"product_id": "123", "quantity": 3}  # Different data
        
        hash1 = IdempotencyManager.compute_request_hash(body1)
        hash2 = IdempotencyManager.compute_request_hash(body2)
        hash3 = IdempotencyManager.compute_request_hash(body3)
        
        # Same data should produce same hash regardless of key order
        assert hash1 == hash2
        
        # Different data should produce different hash
        assert hash1 != hash3


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
