apps/backend/pyproject.toml
[project]
name = "vorte-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.115",
  "uvicorn[standard]>=0.30",
  "motor>=3.6",
  "redis>=5.0",
  "python-jose[cryptography]>=3.3.0",
  "argon2-cffi>=23.1.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-pymongo>=0.47b0",
  "opentelemetry-instrumentation-redis>=0.47b0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-mock>=3.14",
  "httpx>=0.27",
  "ruff>=0.6",
  "faker>=28.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
